+++ lib/widgets/editor.dart
@@ -8,9 +8,11 @@ class Editor extends HookWidget {
   final TextEditingController? controller;
   final FocusNode? focusNode;
   final ValueChanged<String>? onSubmitted;
+  final ValueChanged<String>? onChanged;
   final int? minLines;
   final int? maxLines;
   final String? labelText;
+  final String? initialValue;
   final bool autofocus;
 
   /// Whether the editor should be preview the contents
@@ -22,9 +24,11 @@ class Editor extends HookWidget {
     this.controller,
     this.focusNode,
     this.onSubmitted,
+    this.onChanged,
     this.minLines = 5,
     this.maxLines,
     this.labelText,
+    this.initialValue,
     this.fancy = false,
     required this.instanceHost,
     this.autofocus = false,
@@ -32,7 +36,7 @@ class Editor extends HookWidget {
 
   @override
   Widget build(BuildContext context) {
-    final defaultController = useTextEditingController();
+    final defaultController = useTextEditingController(text: initialValue);
     final actualController = controller ?? defaultController;
 
     if (fancy) {
@@ -51,6 +55,7 @@ class Editor extends HookWidget {
       autofocus: autofocus,
       keyboardType: TextInputType.multiline,
       textCapitalization: TextCapitalization.sentences,
+      onChanged: onChanged,
       onSubmitted: onSubmitted,
       maxLines: maxLines,
       minLines: minLines,